Break down nu-path into separate modules
