+++ crates/nu-path/src/env.rs
@@ -1,0 +1,6 @@
+use std::io;
+use std::path::PathBuf;
+
+pub fn current_dir() -> io::Result<PathBuf> {
+    std::env::current_dir()
+}
